# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/forms": {
            "get": {
                "description": "Возвращает все формы с фильтрацией по статусу и дате формирования",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Формы"
                ],
                "summary": "Получить все формы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "статус формы",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "начальная дата формирования",
                        "name": "formation_date_start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "конечная дата формирвания",
                        "name": "formation_date_end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.AllFormsResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Позволяет изменить комментарий в черновой форме и возвращает обновлённые данные",
                "tags": [
                    "Формы"
                ],
                "summary": "Указать комментарий в форме",
                "parameters": [
                    {
                        "description": "Комментарии",
                        "name": "comments",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemes.UpdateFormRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "description": "Удаляет черновую форму",
                "tags": [
                    "Формы"
                ],
                "summary": "Удалить черновую форму",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/forms/delete_language/{id}": {
            "delete": {
                "description": "Удалить язык программиования из черновой формы",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Формы"
                ],
                "summary": "Удалить язык программирования из черновой формы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id языка программирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/forms/{id}": {
            "get": {
                "description": "Возвращает подробную информацию о форме и комментарий",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Формы"
                ],
                "summary": "Получить одну форму",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id формы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.FormResponse"
                        }
                    }
                }
            }
        },
        "/api/forms/{id}/change_github": {
            "put": {
                "description": "Позволяет изменить ссылку на гитхаб в таблице м-м и возвращает обновленные данные",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Коды"
                ],
                "summary": "Указать ссылку на гитхаб",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id формы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Гитхаб",
                        "name": "github",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemes.ChangeGithubRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.CodeResponse"
                        }
                    }
                }
            }
        },
        "/api/forms/{id}/moderator_confirm": {
            "put": {
                "description": "Подтвердить или отклонить форму модератором",
                "tags": [
                    "Формы"
                ],
                "summary": "Подтвердить/отклонить форму",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id формы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "подтвердить",
                        "name": "confirm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.Language": {
            "type": "object",
            "required": [
                "description",
                "name",
                "subject",
                "task"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "task": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "schemes.AllFormsResponse": {
            "type": "object",
            "properties": {
                "forms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemes.FormOutput"
                    }
                }
            }
        },
...
        "schemes.ChangeGithubRequest": {
            "type": "object",
            "required": [
                "formId",
                "github"
            ],
            "properties": {
                "formId": {
                    "type": "string"
                },
                "github": {
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
        "schemes.CodeOutput": {
            "type": "object",
            "properties": {
                "form_id": {
                    "type": "string"
                },
                "github": {
                    "type": "string"
                },
                "language_id": {
                    "type": "string"
                }
            }
        },
        "schemes.CodeResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/schemes.CodeOutput"
                }
            }
        },
        "schemes.FormOutput": {
            "type": "object",
            "properties": {
                "autotest": {
                    "type": "string"
                },
                "comments": {
                    "type": "string"
                },
                "completion_date": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "formation_date": {
                    "type": "string"
                },
                "moderator": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "student": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "schemes.FormResponse": {
            "type": "object",
            "properties": {
                "form": {
                    "$ref": "#/definitions/schemes.FormOutput"
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Language"
                    }
                }
            }
        },
...
        "schemes.UpdateFormRequest": {
            "type": "object",
            "required": [
                "comments"
            ],
            "properties": {
                "comments": {
                    "type": "string",
                    "maxLength": 300
                }
            }
        }
    }
}